+++ data_pipelines_cli/cli_commands/generate.py
@@ -55,12 +55,17 @@ def _filter_out_dbt_log_output(raw_dbt_output: str) -> str:
     return "\n".join(filter(lambda line: not re.match(pattern, line), raw_dbt_output.splitlines()))
 
 
+def _run_and_filter_dbt_macro(
+    env: str, macro_name: str, macro_args: Dict[str, Any], profiles_path: pathlib.Path
+) -> str:
+    schema_output = _get_unfiltered_macro_run_output(env, macro_name, macro_args, profiles_path)
+    return _filter_out_dbt_log_output(schema_output)
+
+
 def _generate_models_or_sources_from_single_table(
     env: str, macro_name: str, macro_args: Dict[str, Any], profiles_path: pathlib.Path
 ) -> Dict[str, Any]:
-    schema_output = _get_unfiltered_macro_run_output(env, macro_name, macro_args, profiles_path)
-    filtered_output = _filter_out_dbt_log_output(schema_output)
-    return yaml.safe_load(filtered_output)
+    return yaml.safe_load(_run_and_filter_dbt_macro(env, macro_name, macro_args, profiles_path))
 
 
 def _is_ephemeral_model(manifest: Dict[str, Any], model_name: str) -> bool:
@@ -70,10 +75,10 @@ def _is_ephemeral_model(manifest: Dict[str, Any], model_name: str) -> bool:
     raise DataPipelinesError(f"Could not find {model_name} in project's 'manifest.json' file.")
 
 
-def _get_output_yaml_or_warn_if_exists(
-    directory: pathlib.Path, overwrite: bool
+def _get_output_file_or_warn_if_exists(
+    directory: pathlib.Path, overwrite: bool, file_extension: str, filename: Optional[str] = None
 ) -> Optional[pathlib.Path]:
-    output_path = directory.joinpath(f"{directory.name}.yml")
+    output_path = directory.joinpath(f"{filename or directory.name}.{file_extension}")
     if output_path.exists():
         if not overwrite:
             echo_warning(
@@ -96,7 +101,7 @@ def _generate_model_yamls_for_directory(
     macro_arg_name: MacroArgName,
     profiles_path: pathlib.Path,
 ) -> None:
-    output_path = _get_output_yaml_or_warn_if_exists(directory, overwrite)
+    output_path = _get_output_file_or_warn_if_exists(directory, overwrite, "yml")
     if output_path is None:
         return
 
@@ -151,7 +156,7 @@ def generate_source_yamls(
     env: str, source_path: pathlib.Path, overwrite: bool, schema_names: Sequence[str]
 ) -> None:
     profiles_path = generate_profiles_yml(env)
-    output_path = _get_output_yaml_or_warn_if_exists(source_path, overwrite)
+    output_path = _get_output_file_or_warn_if_exists(source_path, overwrite, "yml")
     if output_path is None:
         return
     source_path.mkdir(parents=True, exist_ok=True)
@@ -179,6 +184,43 @@ def generate_source_yamls(
         )
 
 
+def generate_source_sqls(
+    env: str, source_yaml_path: pathlib.Path, staging_path: pathlib.Path, overwrite: bool
+) -> None:
+    profiles_path = generate_profiles_yml(env)
+    staging_path.mkdir(parents=True, exist_ok=True)
+    with open(source_yaml_path, "r") as source_yaml:
+        source_dict = yaml.safe_load(source_yaml)
+        tables_by_source = [
+            (source["name"], table["name"])
+            for source in source_dict["sources"]
+            for table in source["tables"]
+        ]
+    try:
+        for source_name, table_name in tables_by_source:
+            output_path = _get_output_file_or_warn_if_exists(
+                staging_path, overwrite, "sql", f"{table_name}"
+            )
+            if output_path is None:
+                continue
+            table_sql = _run_and_filter_dbt_macro(
+                env,
+                "generate_base_model",
+                {"source_name": source_name, "table_name": table_name},
+                profiles_path,
+            )
+            with open(output_path, "w") as output:
+                output.write(table_sql)
+    except SubprocessNonZeroExitError as err:
+        raise DataPipelinesError(
+            "Error while running dbt command. Ensure that you have codegen "
+            "installed and you have chosen correct existing sources to "
+            "generate table sqls out of.\n" + err.message,
+            submessage=err.submessage,
+        )
+
+
+# -------------------------------- COMMANDS -----------------------------------
 @click.group(name="generate", help="Generate additional dbt files")
 def generate_group() -> None:
     pass
@@ -187,7 +229,7 @@ def generate_group() -> None:
 @generate_group.command(
     name="model-yaml", help="Generate schema YAML using codegen or dbt-profiler"
 )
-@click.option("--env", default="local", type=str, help="Name of the environment")
+@click.option("--env", default="local", type=str, help="Name of the environment", show_default=True)
 @click.option(
     "--with-meta", type=bool, is_flag=True, help="Whether to generate dbt-profiler metadata"
 )
@@ -208,10 +250,11 @@ def _generate_model_yaml_command(
 
 
 @generate_group.command(name="source-yaml", help="Generate source YAML using codegen")
-@click.option("--env", default="local", type=str, help="Name of the environment")
+@click.option("--env", default="local", type=str, help="Name of the environment", show_default=True)
 @click.option(
     "--source-path",
     default=pathlib.Path.cwd().joinpath("models", "source"),
+    show_default=True,
     type=pathlib.Path,
     help="Path to the 'source' directory",
     required=True,
@@ -226,3 +269,32 @@ def _generate_source_command(
     if len(schema_name) == 0:
         raise DataPipelinesError("Command expects at least one 'schema-name' argument")
     generate_source_yamls(env, source_path, overwrite, schema_name)
+
+
+@generate_group.command(
+    name="source-sql", help="Generate SQLs that represents tables in given dataset"
+)
+@click.option("--env", default="local", type=str, help="Name of the environment", show_default=True)
+@click.option(
+    "--source-yaml-path",
+    type=click.Path(exists=True, path_type=pathlib.Path, file_okay=True),
+    default=pathlib.Path.cwd().joinpath("models", "source", "source.yml"),
+    show_default=True,
+    help="Path to the 'source.yml' schema file",
+    required=True,
+)
+@click.option(
+    "--staging-path",
+    default=pathlib.Path.cwd().joinpath("models", "staging"),
+    show_default=True,
+    type=pathlib.Path,
+    help="Path to the 'staging' directory",
+    required=True,
+)
+@click.option(
+    "--overwrite", type=bool, is_flag=True, help="Whether to overwrite existing SQL files"
+)
+def _generate_source_sqls(
+    env: str, source_yaml_path: pathlib.Path, staging_path: pathlib.Path, overwrite: bool
+) -> None:
+    generate_source_sqls(env, source_yaml_path, staging_path, overwrite)